refactor: extract @agents/kg package (Phase 1)
Move 6 KG modules from CLI to packages/kg/:
- graph.ts, graph-schema.ts, graph-lock.ts (graph construction)
- embedder.ts (Ollama embeddings)
- meilisearch.ts (search indexing)
- chunker.ts (markdown chunking + frontmatter)

Move 3 test files (127 tests). Update 5 consumer files.
Leave chunker.ts re-export shim for manifest.ts compat (Phase 2).

KG: 127 pass. CLI: 1557 pass / 10 fail (baseline maintained).
