+++ packages/cli/src/commands/kg.ts
@@ -17,13 +17,8 @@ import { resolve } from 'node:path'
 import { hashFile } from '@agents/core/hash'
 import { currentDir, readText } from '@agents/core/runtime'
 import { type EntityType, EXIT } from '@agents/core/types'
-import { createOutput } from '@agents/sdk/ui'
-import { defineCommand } from 'citty'
-import fg from 'fast-glob'
-import { Ollama } from 'ollama'
-import picomatch from 'picomatch'
-import { chunkMarkdown, parseFrontmatter } from '../lib/chunker'
-import { isOllamaAvailable, prepareEmbeddingText } from '../lib/embedder'
+import { chunkMarkdown, parseFrontmatter } from '@agents/kg/chunker'
+import { isOllamaAvailable, prepareEmbeddingText } from '@agents/kg/embedder'
 import {
   checkHealth,
   createClient,
@@ -35,7 +30,12 @@ import {
   indexEntity,
   searchKeyword,
   searchSemantic,
-} from '../lib/meilisearch'
+} from '@agents/kg/meilisearch'
+import { createOutput } from '@agents/sdk/ui'
+import { defineCommand } from 'citty'
+import fg from 'fast-glob'
+import { Ollama } from 'ollama'
+import picomatch from 'picomatch'
 import { hybridSearch, type RankedResult } from '../lib/search'
 import { globalArgs } from './shared-args'
