+++ packages/cli/src/lib/skill-search-api.ts
@@ -14,9 +14,9 @@
 import { join } from 'node:path'
 import { readText } from '@agents/core/runtime'
 import { CliError, err, ok, type Result } from '@agents/core/types'
+import { checkHealth, createClient, searchKeyword } from '@agents/kg/meilisearch'
 import type { SearchBackendType, SkillSearchResult } from '@agents/sdk/providers/local/schemas'
 import { clampLimit, clampPage } from '@agents/sdk/providers/pagination'
-import { checkHealth, createClient, searchKeyword } from './meilisearch'
 
 // ---------------------------------------------------------------------------
 // Public types
