+++ packages/kg/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@agents/kg",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts",
+    "./*": "./src/*.ts"
+  },
+  "dependencies": {
+    "@agents/core": "workspace:*",
+    "ajv": "^8.18.0",
+    "ajv-formats": "^3.0.1",
+    "graphology": "^0.26.0",
+    "graphology-layout": "^0.6.1",
+    "graphology-layout-forceatlas2": "^0.10.1",
+    "js-yaml": "^4.1.0",
+    "meilisearch": "^0.56.0",
+    "ollama": "^0.5.0"
+  },
+  "devDependencies": {
+    "@types/js-yaml": "^4.0.0"
+  }
+}
