+++ packages/kg/src/graph-lock.ts
