+++ packages/kg/src/graph-schema.ts
