+++ packages/kg/src/graph.ts
